requirements.txt
amqp==5.0.6
asgiref==3.4.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.4
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.7.0
django-dotenv==1.4.2
django-timezone-field==4.2.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.6.3
kombu==5.1.0
Pillow==8.3.0
prompt-toolkit==3.0.19
psycopg2-binary==2.9.3
PyJWT==2.3.0
python-crontab==2.5.1
python-dateutil==2.8.2
pytz==2021.1
redis==3.5.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
typing-extensions==3.10.0.0
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
zipp==3.5.0
